Revert "[FEATURE] make outline message service"
