--- backend/src/main/java/peer/backend/dto/message/LetterTargetDTO.java
@@ -1,13 +1,0 @@
-package peer.backend.dto.message;
-
-import lombok.*;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class LetterTargetDTO {
-    private long targetId;
-    private String targetNickname;
-    private String targetProfile;
-}
