+++ backend/src/main/java/peer/backend/dto/message/MessageUserDTO.java
@@ -1,13 +1,15 @@
 package peer.backend.dto.message;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class SpecificMsgDTO {
-    private long targetId;
-    private long conversationalId;
+public class MessageUserDTO {
+    private String profileImage;
+    private String nickName;
 }
