--- backend/src/main/java/peer/backend/dto/message/MsgDTO.java
@@ -1,17 +1,0 @@
-package peer.backend.dto.message;
-
-import lombok.*;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class MsgDTO {
-    private long senderId;
-    private String senderNickname;
-    private String targetProfile;
-    private long msgId;
-    private String content;
-    private String date;
-    private boolean isEnd;
-}
