--- backend/src/main/java/peer/backend/dto/message/TargetDTO.java
@@ -1,15 +1,0 @@
-package peer.backend.dto.message;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.validation.constraints.NotNull;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class TargetDTO {
-    private long targetId;
-}
