--- backend/src/main/java/peer/backend/dto/messageOld/MessageUserDTO.java
@@ -1,15 +1,0 @@
-package peer.backend.dto.messageOld;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class MessageUserDTO {
-    private String profileImage;
-    private String nickName;
-}
