+++ backend/src/main/java/peer/backend/entity/user/User.java
@@ -11,7 +11,6 @@ import java.util.ArrayList;
 import java.util.List;
 import org.hibernate.annotations.DynamicUpdate;
 import peer.backend.entity.BaseEntity;
-import peer.backend.entity.message.MessageIndex;
 import peer.backend.entity.team.TeamUser;
 import lombok.Setter;
 
@@ -78,10 +77,4 @@ public class User extends BaseEntity {
 
     @OneToMany(mappedBy = "user", cascade = CascadeType.PERSIST, orphanRemoval = true)
     private List<SocialLogin> socialLogins = new ArrayList<>();
-
-    @OneToMany(mappedBy = "user1", cascade = CascadeType.PERSIST)
-    private List<MessageIndex> indexList1 = new ArrayList<>();
-
-    @OneToMany(mappedBy = "user2", cascade = CascadeType.PERSIST)
-    private List<MessageIndex> indexList2 = new ArrayList<>();
 }
